Alternative fix for CVE-2022-4304 (3.0)
Include 2 commits:
1. Revert "Fix Timing Oracle in RSA decryption"
2. Alternative fix for CVE-2022-4304

(Merged from https://github.com/openssl/openssl/pull/20283)
